Refactor and restructure end_model/ to classification/ (clean) (#1216)
* Refactor and restructure end_model/ to classification/

* Refactor Multitask{Dataset, DataLoader} to Classification{}

* Update names of tests

* Delete obsolete test directory

* Remove obsolete snorkel/end_model/ pieces
