+++ snorkel/classification/models/__init__.py
@@ -1,0 +1,2 @@
+from .advanced import AdvancedClassifier, Operation, Task  # noqa: F401
+from .simple_classifier import SimpleClassifier  # noqa: F401
