+++ snorkel/classification/models/advanced/__init__.py
@@ -1,0 +1,2 @@
+from .advanced_classifier import AdvancedClassifier  # noqa: F401
+from .task import Operation, Task  # noqa: F401
