+++ snorkel/classification/training/loggers/checkpointer.py
@@ -6,8 +6,8 @@ from typing import Dict, List
 
 import torch
 
-from snorkel.end_model.snorkel_config import default_config
-from snorkel.end_model.utils import recursive_merge_dicts
+from snorkel.classification.snorkel_config import default_config
+from snorkel.classification.utils import recursive_merge_dicts
 
 
 class Checkpointer(object):
