+++ snorkel/classification/training/schedulers/sequential_scheduler.py
@@ -1,21 +1,20 @@
-from abc import ABC, abstractmethod
 from typing import Any, Dict, Sequence, Tuple
 
-from snorkel.end_model.data import MultitaskDataLoader
+from snorkel.classification.data import ClassifierDataLoader
+
+from .scheduler import Scheduler
 
 
-class Scheduler(ABC):
-    """Return batches from all dataloaders in a specified order
-    """
+class SequentialScheduler(Scheduler):
+    """Return batches from all dataloaders in sequential order."""
 
     def __init__(self):
-        pass
+        super().__init__()
 
-    @abstractmethod
     def get_batches(
-        self, dataloaders: Sequence[MultitaskDataLoader]
-    ) -> Tuple[Tuple[Dict[str, Any], Dict[str, Any]], MultitaskDataLoader]:
-        """Return batches in shuffled order from dataloaders
+        self, dataloaders: Sequence[ClassifierDataLoader]
+    ) -> Tuple[Tuple[Dict[str, Any], Dict[str, Any]], ClassifierDataLoader]:
+        """Return batches from dataloaders sequentially in the order they were given.
 
         Parameters
         ----------
@@ -30,3 +29,6 @@ class Scheduler(ABC):
             model; it is passed primarily so that the model can pull the necessary
             metadata to know what to do with the batch it has been given.
         """
+        for dataloader in dataloaders:
+            for batch in dataloader:
+                yield batch, dataloader
