+++ snorkel/classification/training/schedulers/shuffled_scheduler.py
@@ -1,7 +1,7 @@
 import random
 from typing import Any, Dict, Sequence, Tuple
 
-from snorkel.end_model.data import MultitaskDataLoader
+from snorkel.classification.data import ClassifierDataLoader
 
 from .scheduler import Scheduler
 
@@ -13,8 +13,8 @@ class ShuffledScheduler(Scheduler):
         super().__init__()
 
     def get_batches(
-        self, dataloaders: Sequence[MultitaskDataLoader]
-    ) -> Tuple[Tuple[Dict[str, Any], Dict[str, Any]], MultitaskDataLoader]:
+        self, dataloaders: Sequence[ClassifierDataLoader]
+    ) -> Tuple[Tuple[Dict[str, Any], Dict[str, Any]], ClassifierDataLoader]:
         """Return batches in shuffled order from dataloaders
 
         Parameters
