+++ test/slicing/test_slicing.py
@@ -6,12 +6,11 @@ import pandas as pd
 import torch
 import torch.nn as nn
 
-from snorkel.end_model.data import MultitaskDataLoader, MultitaskDataset
-from snorkel.end_model.model import MultitaskModel
-from snorkel.end_model.modules.utils import ce_loss, softmax
-from snorkel.end_model.scorer import Scorer
-from snorkel.end_model.task import Operation, Task
-from snorkel.end_model.trainer import Trainer
+from snorkel.classification.data import ClassifierDataLoader, ClassifierDataset
+from snorkel.classification.models.advanced import AdvancedClassifier, Operation, Task
+from snorkel.classification.models.advanced.utils import ce_loss, softmax
+from snorkel.classification.scorer import Scorer
+from snorkel.classification.training import Trainer
 from snorkel.slicing.apply import PandasSFApplier
 from snorkel.slicing.sf import slicing_function
 from snorkel.slicing.utils import add_slice_labels, convert_to_slice_tasks
@@ -80,7 +79,7 @@ class SlicingTest(unittest.TestCase):
         # Convert to slice tasks
         task1_tasks = convert_to_slice_tasks(task1, slice_names)
         tasks = task1_tasks + [task2]
-        model = MultitaskModel(tasks=tasks)
+        model = AdvancedClassifier(tasks=tasks)
 
         # Train
         trainer = Trainer(**self.trainer_config)
@@ -109,7 +108,7 @@ def create_dataloader(df, split):
     Y_dict[f"task2_labels"] = torch.LongTensor(df["y2"])
     task_to_label_dict["task2"] = "task2_labels"
 
-    dataset = MultitaskDataset(
+    dataset = ClassifierDataset(
         name="TestData",
         split=split,
         X_dict={
@@ -120,7 +119,7 @@ def create_dataloader(df, split):
         Y_dict=Y_dict,
     )
 
-    dataloader = MultitaskDataLoader(
+    dataloader = ClassifierDataLoader(
         task_to_label_dict=task_to_label_dict,
         dataset=dataset,
         batch_size=4,
